Make uuid to become a module
